+++ src/main/java/main/arbitrage/infrastructure/crawler/UsdToKrwCrawler.java
@@ -11,10 +11,9 @@ import lombok.extern.slf4j.Slf4j;
 public class UsdToKrwCrawler {
     private static final String GOOGLE_FINANCE_USD_TO_KRW_URL =
             "https://www.google.com/finance/quote/USD-KRW";
-    private static final String CURRENT_SELECTOR = "div.YMlKec.fxKbKc"; // 환율 셀렉터 -> 나중에 구글이 수정하면
-                                                                        // 바꿔야함.
+    private static final String CURRENT_SELECTOR = "div.YMlKec.fxKbKc"; // 환율 셀렉터
 
-    public double craw() {
+    public float craw() {
         try {
             Document doc = Jsoup.connect(GOOGLE_FINANCE_USD_TO_KRW_URL).get();
 
@@ -23,7 +22,7 @@ public class UsdToKrwCrawler {
             if (rateToString.isEmpty())
                 return 0;
 
-            return Double.parseDouble(rateToString.replace(",", ""));
+            return Float.parseFloat(rateToString.replace(",", ""));
         } catch (IOException e) {
             // Jsoup로 crawling을 했을때 반환값이 0일경우 아래와 같은 Error Message가 발생함.
             // 에러 확인 후 코드 추가 -> 아직 해당 에러 재발생하지 않음. 추후에 수정.
